Plotting and skipping
A fully functional (i.e., works with the sample dataset) update that allows you to plot your outcome variables (to see if they make sense) and that allows you to skip certain parts (if they have been done in the past). To skip parts of the analysis, the pipeline simply loads previously cleaned data or even spatially aggregated data.